refactor: clean and rename FindBestModelExperiment|
also clean up and docs in other classes
