+++ bunq_ynab_connect/classification/preprocessing/alias_features.py
@@ -17,7 +17,7 @@ class AliasFeatures(Features):
     top_categories: int | None
     enabled: bool
 
-    def __init__(self, top_categories: int | None = None, enabled: bool = True):
+    def __init__(self, top_categories: int | None = None, enabled: bool = True):  # noqa: FBT001, FBT002
         self.top_categories = top_categories
         self.enabled = enabled
